+++ pallets/rc-migrator/src/society.rs
@@ -528,7 +528,7 @@ impl<T: Config> PalletMigration for SocietyMigrator<T> {
 			}
 
 			if T::MaxAhWeight::get()
-				.any_lt(Self::receive_society_messages_weight((messages.len() + 1) as u32))
+				.any_lt(Self::receive_society_messages_weight(messages.len() + 1))
 			{
 				log::info!(
 					target: LOG_TARGET,
